wip on feed logic... nearly there.
