+++ src/App.js
@@ -4,7 +4,7 @@ import HicetnuncContextProvider from './context/HicetnuncContext'
 import Sync from './pages/sync'
 import About from './pages/about'
 import Display from './pages/display'
-import Feed from './pages/feed'
+import { Feed } from './pages/feed'
 import Mint from './pages/mint'
 import ObjktDisplay from './pages/objkt-display'
 import Loading from './pages/loading'
@@ -12,14 +12,13 @@ import { Header } from './components/header'
 import { Footer } from './components/footer'
 import './App.css'
 
-function App() {
+const App = () => {
   return (
     <HicetnuncContextProvider>
       <Header />
       <Switch>
         <Route exact path="/" component={Feed} />
         <Route path="/tz/:id" component={Display} />
-        <Route path="/feed" component={Feed} />
         <Route path="/about" component={About} />
         <Route path="/sync" component={Sync} />
         <Route path="/mint" component={Mint} />
