+++ src/api/index.js
@@ -1,6 +1,20 @@
 const axios = require('axios')
 
 /**
+ * Gets Feed for homepage
+ */
+export const GetFeed = async ({ counter }) => {
+  return new Promise((resolve, reject) => {
+    axios
+      .post(process.env.REACT_APP_FEED, { counter: counter })
+      .then((res) => {
+        resolve(res.data.result)
+      })
+      .catch((e) => reject(e))
+  })
+}
+
+/**
  * Get OBJKT detail page
  */
 export const GetOBJKT = async ({ objkt_id }) => {
