+++ src/components/feed-item/index.module.scss
@@ -1,0 +1,3 @@
+.container {
+  border: 1px dashed red;
+}
