+++ src/components/layout/page/index.module.scss
@@ -2,6 +2,7 @@
 @import '../../../styles/layout.scss';
 
 .container {
+  position: relative;
   width: 100%;
   min-height: 100vh;
   padding: 100px 0;
